refactor: rename jobs dir to timer
