+++ uaclient/entitlements/tests/test_esm.py
@@ -15,7 +15,7 @@ def entitlement(request, entitlement_factory):
     return entitlement_factory(request.param, suites=["xenial"])
 
 
-@mock.patch("uaclient.jobs.update_messaging.update_motd_messages")
+@mock.patch("uaclient.timer.update_messaging.update_motd_messages")
 @mock.patch(
     "uaclient.system.get_release_info",
     return_value=mock.MagicMock(series="xenial"),
