+++ uaclient/timer/update_contract_info.py
