+++ uaclient/timer/update_messaging.py
